Trait finchers::action::EndpointAction

pub trait EndpointAction<Bd> {
    type Output: Tuple;
    fn poll_action(
        &mut self, 
        cx: &mut ActionContext<Bd>
    ) -> Poll<Self::Output, Error>;

    fn preflight(
        &mut self, 
        cx: &mut PreflightContext
    ) -> Result<Preflight<Self::Output>, Error> { ... }
}

A trait that abstracts the action of endpoints.

Associated Types

type Output: Tuple

The type returned from this action.

Required Methods

fn poll_action(
&mut self,
cx: &mut ActionContext<Bd>
) -> Poll<Self::Output, Error>

Progress this action and returns the result if ready.

Provided Methods

fn preflight(
&mut self,
cx: &mut PreflightContext
) -> Result<Preflight<Self::Output>, Error>

Applies an incoming request to this action and returns the result if possible.

This method is called only once after the instance of Self is created. The signature of this method is almost the same as poll_action, but there are the following differences:

The return type of this method is not Poll<Self::Output, Self::Error>. When this method returns an Ok(Incomplete), it means that the action needs to poll the internal resource or access ActionContext to complete itself. Therefore, it is not possible to poll any asynchronous objects that this action holds within this method.

The result of this method may affect the behavior of routing. If this method will return an Err, it is possible that the combinator may ignore this error value and choose another EndpointAction without aborting the process (in contrast, the error values returned from poll_action are not ignored).

Some limitations are added to PreflightContext in order to keep consistency when another endpoint returns an error (for example, it cannot be taken the instance of request body inside of this method).

By default, this method does nothing and immediately returns an Ok(Preflight::Incomplete).