Module finchers::endpoint

Components for constructing Endpoint.

Re-exports

pub use self::context::with_get_cx;

pub use self::context::ApplyContext;

pub use self::context::TaskContext;

pub use self::error::ApplyError;

pub use self::error::ApplyResult;

pub use self::wrapper::EndpointWrapExt;

pub use self::wrapper::Wrapper;

Modules

context
The definition of contextual information during applying endpoints.

error
Definition of ApplyError and supplemental components.

syntax
Components for building endpoints which matches to a specific HTTP path.

wrapper
Built-in wrappers.

Structs

And

ApplyFn

ByRef

Cloned

EndpointObj

LocalEndpointObj

Or

OrStrict

SendEndpoint
A wrapper struct which wraps a value whose type implements IsSendEndpoint and provides the implementations of Endpoint<'a>.

Unit

Traits

Endpoint
Trait representing an endpoint.

IntoEndpoint
Trait representing the transformation into an Endpoint.

IntoEndpointExt
A set of extension methods for composing multiple endpoints.

IsSendEndpoint
A trait representing an endpoint with a constraint that the returned "Future" to be transferred across thread boundaries.

Functions

apply_fn
Create an endpoint from a function.

by_ref
Create an endpoint which simply returns a reference to the specified value.

cloned
Create an endpoint which simply clones the specified value.

unit
Create an endpoint which simply returns an unit (()).