Struct finchers::endpoint::error::ApplyError

pub struct ApplyError(_);

A type representing error values returned from Endpoint::apply().

This error type represents the errors around routing determined before executing the Future returned from the endpoint.

Methods

impl ApplyError

pub fn not_matched() -> ApplyError

Create a value of ApplyError with an annotation that the current endpoint does not match to the provided request.

pub fn method_not_allowed(allowed: Verbs) -> ApplyError

Create a value of ApplyError with an annotation that the current endpoint does not matche to the provided HTTP method.

pub fn custom(
err: impl Into<Error>
) -> ApplyError

Create a value of ApplyError from the custom error value.

The generated error has the HTTP status code 400 Bad Request.