Trait finchers::endpoint::Endpoint

pub trait Endpoint<Bd>: IsEndpoint {
    type Output: Tuple;
    type Action: EndpointAction<Bd, Output = Self::Output>;
    fn action(&self) -> Self::Action;

    fn with_output<T: Tuple>(self) -> Self
    where
        Self: Endpoint<Bd, Output = T> + Sized,
    { ... }
}

Trait representing an endpoint, the main trait for abstracting HTTP services in Finchers.

The endpoint behaves as an asynchronous process that takes an HTTP request and convert it into a value of the associated type. The process that handles an request is abstracted by EndpointAction, and that instances are constructed by Endpoint for each request.

Associated Types

type Output: Tuple

The type of value that will be returned from Action.

type Action: EndpointAction<Bd, Output = Self::Output>

The type of EndpointAction associated with this endpoint.

Required Methods

fn action(&self) -> Self::Action

Spawns an instance of Action that applies to an request.

Provided Methods

fn with_output<T: Tuple>(self) -> Self where
Self: Endpoint<Bd, Output = T> + Sized,

Add an annotation that the associated type Output is fixed to T.