Trait finchers::endpoint::IsEndpoint

pub trait IsEndpoint {
    fn boxed<Bd, T>(self) -> EndpointObj<Bd, T>
    where
        Self: Endpoint<Bd, Output = T> + Send + Sync + 'static + Sized,
        Self::Action: Send + 'static,
        T: Tuple,
    { ... }

    fn boxed_local<Bd, T>(self) -> LocalEndpointObj<Bd, T>
    where
        Self: Endpoint<Bd, Output = T> + 'static + Sized,
        Self::Action: 'static,
        T: Tuple,
    { ... }
}

A trait indicating that the type has an implementation of Endpoint<Bd>.

The purpose of this trait is to implement the extension methods to Endpoints in situation when the type of request body is unknown.

Provided Methods

fn boxed<Bd, T>(self) -> EndpointObj<Bd, T> where
Self: Endpoint<Bd, Output = T> + Send + Sync + 'static + Sized,
Self::Action: Send + 'static,
T: Tuple,

Converts this endpoint into an EndpointObj.

fn boxed_local<Bd, T>(self) -> LocalEndpointObj<Bd, T> where
Self: Endpoint<Bd, Output = T> + 'static + Sized,
Self::Action: 'static,
T: Tuple,

Converts this endpoint into a LocalEndpointObj.