Struct finchers::error::Error

pub struct Error(_);

A type which holds a value of HttpError in a type-erased form.

Methods

impl Error

pub fn is<T: HttpError>(&self) -> bool

Returns true if the type of contained value is the same as T.

pub fn downcast_ref<T: HttpError>(&self) -> Option<&T>

Attempts to downcast the boxed value to a conrete type by reference.

pub fn downcast_mut<T: HttpError>(&mut self) -> Option<&mut T>

Attempts to downcast the boxed value to a conrete type by mutable reference.

pub fn downcast<T: HttpError>(self) -> Result<T, Error>

Attempts to downcast the boxed value to a conrete type.

pub fn status_code(&self) -> StatusCode

Return the HTTP status code associated with contained value.

pub fn headers(&self, headers: &mut HeaderMap)

Append a set of header values to the header map.

pub fn cause(&self) -> Option<&Fail>

Returns a reference to the underlying cause of contained error value.