Trait finchers::error::HttpError

pub trait HttpError: AsFail + Send + Sync + 'static {
    fn status_code(&self) -> StatusCode { ... }

    fn to_response(&self, _: &Request<()>) -> Response<()> { ... }
}

Trait that abstracts the error representation used in Finchers.

Roughly speaking, this trait adds some context around HTTP to failure::Fail.

Provided Methods

fn status_code(&self) -> StatusCode

Returns an HTTP status code associated with this error value.

fn to_response(&self, _: &Request<()>) -> Response<()>

Creates an HTTP response without the request body.

Methods

impl dyn HttpError

pub fn is<T: HttpError>(&self) -> bool

Returns true if the type of contained value is the same as T.

pub fn downcast_ref<T: HttpError>(&self) -> Option<&T>

Attempts to downcast the boxed value to a conrete type by reference.

pub fn downcast_mut<T: HttpError>(&mut self) -> Option<&mut T>

Attempts to downcast the boxed value to a conrete type by mutable reference.

pub fn downcast<T: HttpError>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Attempts to downcast the boxed value to a conrete type.