Module finchers::input::cookie

Components for managing Cookie values.

Structs

Cookie
Representation of an HTTP cookie.

CookieBuilder
Structure that follows the builder pattern for building Cookie structs.

Cookies
A proxy object for tracking changes of Cookie values.

Duration
ISO 8601 time duration with nanosecond precision. This also allows for the negative duration; see individual methods for details.

EncodedCookie
Wrapper around Cookie whose Display implementation percent-encodes the cookie's name and value.

Tm
Holds a calendar date and time broken down into its components (year, month, day, and so on), also called a broken-down time value.

Enums

SameSite
The SameSite cookie attribute.