Struct finchers::input::Input

pub struct Input { /* fields omitted */ }

The contextual information with an incoming HTTP request.

Methods

impl Input

pub fn request(&self) -> &Request<ReqBody>

Return a shared reference to the value of raw HTTP request without the message body.

pub fn payload(&mut self) -> Option<Body>

Takes the instance of RequestBody from this value.

pub fn content_type(&mut self) -> Result<Option<&Mime>, Error>

Attempts to get the entry of Content-type and parse its value.

The result of this method is cached and it will return the reference to the cached value on subsequent calls.

pub fn cookies(&mut self) -> Result<&mut CookieJar, Error>

Returns a Cookies<'_> or initialize the internal Cookie jar.

pub fn response_headers(&mut self) -> &mut HeaderMap

Returns a mutable reference to a HeaderMap which contains the entries of response headers.

The values inserted in this header map are automatically added to the actual response.