Struct finchers::local::LocalRequest

pub struct LocalRequest { /* fields omitted */ }

A builder of dummy HTTP request.

Methods

impl LocalRequest

pub fn method<M>(self, method: M) -> Self where
Method: HttpTryFrom<M>,

Overwrite the HTTP method of this dummy request with given value.

Panics

This method will panic if the parameter is invalid HTTP method.

pub fn uri<U>(self, uri: U) -> Self where
Uri: HttpTryFrom<U>,

Overwrite the URI of this dummy request with given value.

Panics

This method will panic if the parameter is invalid HTTP method.

pub fn header<K, V>(self, name: K, value: V) -> Self where
HeaderName: HttpTryFrom<K>,
HeaderValue: HttpTryFrom<V>,

Append the given header entry into this dummy request.

Panics

This method will panic if the given header name or value is invalid.

pub fn body(
self,
body: impl Into<Body>
) -> Self

Overwrite the message body of this dummy request with given instance.

pub fn apply<'e, E: Endpoint<'e>>(
self,
endpoint: &'e E
) -> Result<E::Output, Error>

Apply this dummy request to the associated endpoint and get its response.

pub fn respond<'e, E>(self, endpoint: &'e E) -> Response<ResBody> where
E: Endpoint<'e>,
E::Output: Output,