Trait finchers::output::body::ResBody

pub trait ResBody {
    type Data: Buf + Send;
    type Error: Into<Box<Error + Send + Sync + 'static>>;
    type Payload: Payload<Data = Self::Data, Error = Self::Error>;
    fn into_payload(self) -> Self::Payload;
}

The trait representing the message body in an HTTP response.

Associated Types

type Data: Buf + Send

A buffer of bytes representing a single chunk of a message body.

type Error: Into<Box<Error + Send + Sync + 'static>>

The error type of Self::Payload.

type Payload: Payload<Data = Self::Data, Error = Self::Error>

The type of Payload to be convert from itself.

Required Methods

fn into_payload(self) -> Self::Payload

Converts itself into a Payload.