Struct finchers::test::TestRunner

pub struct TestRunner<E> { /* fields omitted */ }

A test runner for emulating the behavior of endpoints in the server.

It uses internally the current thread version of Tokio runtime for executing asynchronous processes.

Methods

impl<E> TestRunner<E> where
E: Endpoint<ReqBody>,

pub fn new(endpoint: E) -> Result<TestRunner<E>>

Create a TestRunner from the specified endpoint.

pub fn with_runtime(endpoint: E, rt: Runtime) -> TestRunner<E>

Create a TestRunner from the specified endpoint with a Tokio runtime.

pub fn default_headers(&mut self) -> &mut HeaderMap

Returns a reference to the header map, whose values are set before applying the request to endpoint.

pub fn endpoint(&mut self) -> &mut E

Returns a reference to the instance of Endpoint owned by this runner.

pub fn runtime(&mut self) -> &mut Runtime

Returns a reference to the Tokio runtime managed by this runner.

pub fn apply<T>(
&mut self,
request: impl TestRequest
) -> Result<T, Error> where
E: Endpoint<ReqBody, Output = (T,)>,

Applies the given request to the inner endpoint and retrieves the result of returned future.

This method is available only if the output of endpoint is a tuple with a single element. If the output type is an unit or the tuple contains more than one element, use apply_raw instead.

pub fn apply_raw(
&mut self,
request: impl TestRequest
) -> Result<E::Output, Error>

Applies the given request to the inner endpoint and retrieves the result of returned future without peeling tuples.