Trait finchers::endpoint::Endpoint

pub trait Endpoint<'a>: 'a {
    type Output: Tuple;
    type Future: TryFuture<Ok = Self::Output, Error = Error> + 'a;
    fn apply(&'a self, ecx: &mut Context) -> EndpointResult<Self::Future>;

    fn with_output<T: Tuple>(self) -> Self
    where
        Self: Endpoint<'a, Output = T> + Sized,
    { ... }

    fn wrap<W>(self, wrapper: W) -> W::Endpoint
    where
        Self: Sized,
        W: Wrapper<'a, Self>,
    { ... }
}

Trait representing an endpoint.

Associated Types

type Output: Tuple

The inner type associated with this endpoint.

type Future: TryFuture<Ok = Self::Output, Error = Error> + 'a

The type of value which will be returned from apply.

Required Methods

fn apply(&'a self, ecx: &mut Context) -> EndpointResult<Self::Future>

Perform checking the incoming HTTP request and returns an instance of the associated Future if matched.

Provided Methods

fn with_output<T: Tuple>(self) -> Self where
Self: Endpoint<'a, Output = T> + Sized,

Add an annotation that the associated type Output is fixed to T.

fn wrap<W>(self, wrapper: W) -> W::Endpoint where
Self: Sized,
W: Wrapper<'a, Self>,

Converts self using the provided Wrapper.