Trait finchers::error::HttpError

pub trait HttpError: Debug + Display + Send + Sync + 'static {
    fn status_code(&self) -> StatusCode { ... }

    fn headers(&self, headers: &mut HeaderMap) { ... }

    fn cause(&self) -> Option<&Fail> { ... }
}

Trait representing error values from endpoints.

The types which implements this trait will be implicitly converted to an HTTP response by the runtime.

Provided Methods

fn status_code(&self) -> StatusCode

Return the HTTP status code associated with this error type.

fn headers(&self, headers: &mut HeaderMap)

Append a set of header values to the header map.

fn cause(&self) -> Option<&Fail>